src/stitches-py/requirements.txt
typing-inspect
jinja2
pydantic
asyncclick
asyncio